refactor : DefualtFireBaseRepository 파일 별도로 분리
